`/assessments`: split out separate `/admin` endpoint
