+++ test/cadet_web/controllers/assessments_controller_test.exs
@@ -3,14 +3,12 @@ defmodule CadetWeb.AssessmentsControllerTest do
   use Timex
 
   import Ecto.Query
-  import ExUnit.CaptureLog
   import Mock
 
   alias Cadet.{Assessments, Repo}
   alias Cadet.Accounts.{Role, User}
   alias Cadet.Assessments.{Assessment, Submission, SubmissionStatus}
   alias Cadet.Autograder.GradingJob
-  alias Cadet.Test.XMLGenerator
   alias CadetWeb.AssessmentsController
 
   @local_name "test/fixtures/local_repo"
@@ -1175,372 +1173,9 @@ defmodule CadetWeb.AssessmentsControllerTest do
     end
   end
 
-  describe "POST /, unauthenticated" do
-    test "unauthorized", %{conn: conn} do
-      assessment = build(:assessment, type: "mission", is_published: true)
-      questions = build_list(5, :question, assessment: nil)
-      xml = XMLGenerator.generate_xml_for(assessment, questions)
-      file = File.write("test/fixtures/local_repo/test.xml", xml)
-      force_update = "false"
-      body = %{assessment: file, forceUpdate: force_update}
-      conn = post(conn, build_url(), body)
-      assert response(conn, 401) =~ "Unauthorised"
-    end
-  end
-
-  describe "POST /, student only" do
-    @tag authenticate: :student
-    test "unauthorized", %{conn: conn} do
-      assessment = build(:assessment, type: "mission", is_published: true)
-      questions = build_list(5, :question, assessment: nil)
-      xml = XMLGenerator.generate_xml_for(assessment, questions)
-      force_update = "false"
-      body = %{assessment: xml, forceUpdate: force_update}
-      conn = post(conn, build_url(), body)
-      assert response(conn, 403) == "User not allowed to create assessment"
-    end
-  end
-
-  describe "POST /, staff only" do
-    @tag authenticate: :staff
-    test "successful", %{conn: conn} do
-      assessment = build(:assessment, type: "mission", is_published: true)
-      questions = build_list(5, :question, assessment: nil)
-
-      xml = XMLGenerator.generate_xml_for(assessment, questions)
-      force_update = "false"
-      path = Path.join(@local_name, "connTest")
-      file_name = "test.xml"
-      location = Path.join(path, file_name)
-      File.mkdir_p!(path)
-      File.write!(location, xml)
-
-      formdata = %Plug.Upload{
-        content_type: "text/xml",
-        filename: file_name,
-        path: location
-      }
-
-      body = %{assessment: %{file: formdata}, forceUpdate: force_update}
-      conn = post(conn, build_url(), body)
-      number = assessment.number
-
-      expected_assessment =
-        Assessment
-        |> where(number: ^number)
-        |> Repo.one()
-
-      assert response(conn, 200) == "OK"
-      assert expected_assessment != nil
-    end
-
-    @tag authenticate: :staff
-    test "upload empty xml", %{conn: conn} do
-      xml = ""
-      force_update = "true"
-      path = Path.join(@local_name, "connTest")
-      file_name = "test.xml"
-      location = Path.join(path, file_name)
-      File.mkdir_p!(path)
-      File.write!(location, xml)
-
-      formdata = %Plug.Upload{
-        content_type: "text/xml",
-        filename: file_name,
-        path: location
-      }
-
-      body = %{assessment: %{file: formdata}, forceUpdate: force_update}
-
-      err_msg =
-        "Invalid XML fatal expected_element_start_tag file file_name_unknown line 1 col 1 "
-
-      assert capture_log(fn ->
-               conn = post(conn, build_url(), body)
-               assert(response(conn, 400) == err_msg)
-             end) =~ ~r/.*fatal: :expected_element_start_tag.*/
-    end
-  end
-
-  describe "DELETE /:assessment_id, unauthenticated" do
-    test "unauthorized", %{conn: conn} do
-      assessment = insert(:assessment)
-      conn = delete(conn, build_url(assessment.id))
-      assert response(conn, 401) =~ "Unauthorised"
-    end
-  end
-
-  describe "DELETE /:assessment_id, student only" do
-    @tag authenticate: :student
-    test "unauthorized", %{conn: conn} do
-      assessment = insert(:assessment)
-      conn = delete(conn, build_url(assessment.id))
-      assert response(conn, 403) == "User is not permitted to delete"
-    end
-  end
-
-  describe "DELETE /:assessment_id, staff only" do
-    @tag authenticate: :staff
-    test "successful", %{conn: conn} do
-      assessment = insert(:assessment)
-      conn = delete(conn, build_url(assessment.id))
-      assert response(conn, 200) == "OK"
-      assert Repo.get(Assessment, assessment.id) == nil
-    end
-  end
-
-  describe "POST /publish/:assessment_id, unauthenticated" do
-    test "unauthorized", %{conn: conn} do
-      assessment = insert(:assessment)
-      conn = post(conn, build_url_publish(assessment.id))
-      assert response(conn, 401) =~ "Unauthorised"
-    end
-  end
-
-  describe "POST /publish/:assessment_id, student only" do
-    @tag authenticate: :student
-    test "forbidden", %{conn: conn} do
-      assessment = insert(:assessment)
-      conn = post(conn, build_url_publish(assessment.id))
-      assert response(conn, 403) == "User is not permitted to publish"
-    end
-  end
-
-  describe "POST /publish/:assessment_id, staff only" do
-    @tag authenticate: :staff
-    test "successful toggle from published to unpublished", %{conn: conn} do
-      assessment = insert(:assessment, is_published: true)
-      conn = post(conn, build_url_publish(assessment.id))
-      expected = Repo.get(Assessment, assessment.id).is_published
-      assert response(conn, 200) == "OK"
-      assert expected == false
-    end
-
-    @tag authenticate: :staff
-    test "successful toggle from unpublished to published", %{conn: conn} do
-      assessment = insert(:assessment, is_published: false)
-      conn = post(conn, build_url_publish(assessment.id))
-      expected = Repo.get(Assessment, assessment.id).is_published
-      assert response(conn, 200) == "OK"
-      assert expected == true
-    end
-  end
-
-  describe "POST /update/:assessment_id, unauthenticated" do
-    test "unauthorized", %{conn: conn} do
-      assessment = insert(:assessment)
-      conn = post(conn, build_url_update(assessment.id))
-      assert response(conn, 401) =~ "Unauthorised"
-    end
-  end
-
-  describe "POST /update/:assessment_id, student only" do
-    @tag authenticate: :student
-    test "forbidden", %{conn: conn} do
-      new_open_at =
-        Timex.now()
-        |> Timex.beginning_of_day()
-        |> Timex.shift(days: 3)
-        |> Timex.shift(hours: 4)
-
-      new_open_at_string =
-        new_open_at
-        |> Timex.format!("{ISO:Extended}")
-
-      new_close_at = Timex.shift(new_open_at, days: 7)
-
-      new_close_at_string =
-        new_close_at
-        |> Timex.format!("{ISO:Extended}")
-
-      new_dates = %{openAt: new_open_at_string, closeAt: new_close_at_string}
-      assessment = insert(:assessment)
-      conn = post(conn, build_url_update(assessment.id), new_dates)
-      assert response(conn, 403) == "User is not permitted to edit"
-    end
-  end
-
-  describe "POST /update/:assessment_id, staff only" do
-    @tag authenticate: :staff
-    test "successful", %{conn: conn} do
-      open_at =
-        Timex.now()
-        |> Timex.beginning_of_day()
-        |> Timex.shift(days: 3)
-        |> Timex.shift(hours: 4)
-
-      close_at = Timex.shift(open_at, days: 7)
-      assessment = insert(:assessment, %{open_at: open_at, close_at: close_at})
-
-      new_open_at =
-        open_at
-        |> Timex.shift(days: 3)
-
-      new_open_at_string =
-        new_open_at
-        |> Timex.format!("{ISO:Extended}")
-
-      new_close_at =
-        close_at
-        |> Timex.shift(days: 5)
-
-      new_close_at_string =
-        new_close_at
-        |> Timex.format!("{ISO:Extended}")
-
-      new_dates = %{openAt: new_open_at_string, closeAt: new_close_at_string}
-
-      conn =
-        conn
-        |> post(build_url_update(assessment.id), new_dates)
-
-      assessment = Repo.get(Assessment, assessment.id)
-      assert response(conn, 200) == "OK"
-      assert [assessment.open_at, assessment.close_at] == [new_open_at, new_close_at]
-    end
-
-    @tag authenticate: :staff
-    test "allowed to change open time of opened assessments", %{conn: conn} do
-      open_at =
-        Timex.now()
-        |> Timex.beginning_of_day()
-        |> Timex.shift(days: -3)
-        |> Timex.shift(hours: 4)
-
-      close_at = Timex.shift(open_at, days: 7)
-      assessment = insert(:assessment, %{open_at: open_at, close_at: close_at})
-
-      new_open_at =
-        open_at
-        |> Timex.shift(days: 6)
-
-      new_open_at_string =
-        new_open_at
-        |> Timex.format!("{ISO:Extended}")
-
-      close_at_string =
-        close_at
-        |> Timex.format!("{ISO:Extended}")
-
-      new_dates = %{openAt: new_open_at_string, closeAt: close_at_string}
-
-      conn =
-        conn
-        |> post(build_url_update(assessment.id), new_dates)
-
-      assessment = Repo.get(Assessment, assessment.id)
-      assert response(conn, 200) == "OK"
-      assert [assessment.open_at, assessment.close_at] == [new_open_at, close_at]
-    end
-
-    @tag authenticate: :staff
-    test "not allowed to set close time to before open time", %{conn: conn} do
-      open_at =
-        Timex.now()
-        |> Timex.beginning_of_day()
-        |> Timex.shift(days: 3)
-        |> Timex.shift(hours: 4)
-
-      close_at = Timex.shift(open_at, days: 7)
-      assessment = insert(:assessment, %{open_at: open_at, close_at: close_at})
-
-      new_close_at =
-        open_at
-        |> Timex.shift(days: -1)
-
-      new_close_at_string =
-        new_close_at
-        |> Timex.format!("{ISO:Extended}")
-
-      open_at_string =
-        open_at
-        |> Timex.format!("{ISO:Extended}")
-
-      new_dates = %{openAt: open_at_string, closeAt: new_close_at_string}
-
-      conn =
-        conn
-        |> post(build_url_update(assessment.id), new_dates)
-
-      assessment = Repo.get(Assessment, assessment.id)
-      assert response(conn, 400) == "New end date should occur after new opening date"
-      assert [assessment.open_at, assessment.close_at] == [open_at, close_at]
-    end
-
-    @tag authenticate: :staff
-    test "successful, set close time to before current time", %{conn: conn} do
-      open_at =
-        Timex.now()
-        |> Timex.beginning_of_day()
-        |> Timex.shift(days: -3)
-        |> Timex.shift(hours: 4)
-
-      close_at = Timex.shift(open_at, days: 7)
-      assessment = insert(:assessment, %{open_at: open_at, close_at: close_at})
-
-      new_close_at =
-        Timex.now()
-        |> Timex.shift(days: -1)
-
-      new_close_at_string =
-        new_close_at
-        |> Timex.format!("{ISO:Extended}")
-
-      open_at_string =
-        open_at
-        |> Timex.format!("{ISO:Extended}")
-
-      new_dates = %{openAt: open_at_string, closeAt: new_close_at_string}
-
-      conn =
-        conn
-        |> post(build_url_update(assessment.id), new_dates)
-
-      assessment = Repo.get(Assessment, assessment.id)
-      assert response(conn, 200) == "OK"
-      assert [assessment.open_at, assessment.close_at] == [open_at, new_close_at]
-    end
-
-    @tag authenticate: :staff
-    test "successful, set open time to before current time", %{conn: conn} do
-      open_at =
-        Timex.now()
-        |> Timex.beginning_of_day()
-        |> Timex.shift(days: 3)
-        |> Timex.shift(hours: 4)
-
-      close_at = Timex.shift(open_at, days: 7)
-      assessment = insert(:assessment, %{open_at: open_at, close_at: close_at})
-
-      new_open_at =
-        Timex.now()
-        |> Timex.shift(days: -1)
-
-      new_open_at_string =
-        new_open_at
-        |> Timex.format!("{ISO:Extended}")
-
-      close_at_string =
-        close_at
-        |> Timex.format!("{ISO:Extended}")
-
-      new_dates = %{openAt: new_open_at_string, closeAt: close_at_string}
-
-      conn =
-        conn
-        |> post(build_url_update(assessment.id), new_dates)
-
-      assessment = Repo.get(Assessment, assessment.id)
-      assert response(conn, 200) == "OK"
-      assert [assessment.open_at, assessment.close_at] == [new_open_at, close_at]
-    end
-  end
-
-  defp build_url, do: "/v1/assessments/"
-  defp build_url(assessment_id), do: "/v1/assessments/#{assessment_id}"
-  defp build_url_submit(assessment_id), do: "/v1/assessments/#{assessment_id}/submit"
-  defp build_url_publish(assessment_id), do: "/v1/assessments/publish/#{assessment_id}"
-  defp build_url_update(assessment_id), do: "/v1/assessments/update/#{assessment_id}"
+  defp build_url, do: "/v2/assessments/"
+  defp build_url(assessment_id), do: "/v2/assessments/#{assessment_id}"
+  defp build_url_submit(assessment_id), do: "/v2/assessments/#{assessment_id}/submit"
 
   defp open_at_asc_comparator(x, y), do: Timex.before?(x.open_at, y.open_at)
